Crate problem104

Problem 104 (Pandigital Fibonacci ends)

Solution detail

Finding all Fibonacci numbers which are pandigital in the last 9 digits is fairly easy - simply calculate the Fibonacci sequence using the usual iterative method, but only storing the terms modulo 10^9.

The real problem is, given an index k for which Fₖ is pandigital in the last 9 digits, checking whether it is also pandigital in the first 9 digits. We surely cannot calculate the full number, because the number of digits would be enormous.

Instead, we can do some maths with logarithms to work out an approximation to Fₖ which is good enough to be used to calculate the first 9 digits.

Specifically, since for large k it is true that Fₖ is approximately equal to ϕᵏ / √5, where ϕ is the golden ratio, we have the relationship:

First 9 digits of Fₖ ≈ 10^({ k * log₁₀(ϕ) - log₁₀(√5) }) * 10⁸

In the above, the braces {} represent the fractional part. This is fast and accurate enough for our purposes.

Structs

FibonacciIterator
Iterate over the Fibonacci numbers, modulo 10^9

Functions

answer
Solve the problem, returning the answer as a String

is_fibonacci_left_pandigital
Check whether the kth Fibonacci number is pandigital in the first 9 digits.

is_pandigital
Check whether the given number is pandigital.

main

solve
Find the index of the first Fibonacci number which is pandigital at both ends.